requirements.txt
twisted
requests
py-august